Plot annual cycle of min/max daily temperatures.
Plots takes CSV file, does min/max division for each day and plots the result.
NOTE: 
Try files excluding 2018's. Program skips number of rows (as they are just comments) and apparently 2018 files are not similar to earlier ones.
I tried St. John's int A for years 2012 and 2017 and it was okay. if there was and error about this, you only need to change number of skipped lines.